+++ modules/monitoring_function/d-myip.tf
@@ -1,0 +1,3 @@
+data "http" "myip" {
+  url = "http://ip4.clara.net/?raw"
+}