+++ modules/monitoring_function/versions.tf
@@ -10,5 +10,9 @@ terraform {
       source  = "aztfmod/azurecaf"
       version = "~> 1.1"
     }
+    http = {
+      source  = "hashicorp/http"
+      version = ">= 2.0"
+    }
   }
 }